Re-enable pagebuilder/selections test
