+++ tests/cypress/page-object/jcontent.ts
@@ -410,7 +410,11 @@ class PageBuilderModule extends BaseComponent {
     }
 
     contextMenu(selectFirst = false, force = true): Menu {
-        this.getHeader(selectFirst);
+        if (selectFirst) {
+            this.getHeader(selectFirst);
+        } else {
+            this.hover();
+        }
         this.get().rightclick({force});
         return getComponentBySelector(Menu, '#menuHolder .moonstone-menu:not(.moonstone-hidden)');
     }
